bus in override equals

+++ app/src/main/java/com/clinic/anhe/medicinetracker/model/MedicineRecordCardViewModel.java
@@ -117,7 +117,7 @@ public class MedicineRecordCardViewModel {
 
         MedicineRecordCardViewModel record = (MedicineRecordCardViewModel) obj;
 
-        return record.getRid()==rid;
+        return record.getRid().intValue()==rid.intValue();
     }
 
     @Override
